MARBLE-876 Bugfix Browse page with url hash
Reported problem:
Going to `/browse` works fine, but going to `/browse#date` caused page 
to become unresponsive and links would not work.

Observed problem:
Going to `/browse#date` caused page to hang and an infinite loop was 
triggered by Okta Auth library being unable to read the cookie.

Actual problem and soloution:
Due to the way Gatsby and Reach Router handle urls internally, there was 
a conflict with the Okta Auth library causing the Okta Auth library to 
recieve incorrect information about the referring url. Since hash urls 
can be problematic in SPA and Gatsby apps and are rarely used in our 
current setup, I've opted to replace links of the format `/browse#date` 
with `/browse?scrollto=date` and add a conditional check on the `Layout` 
component to handle targeting and scrolling to the correct region of the 
page.